requirements.txt
numpy
matplotlib
pandas
openpyxl
plotly_express
streamlit
seaborn
bokeh